This MR can make it work in vs2015 (#26)
* Make CMakeList.txt work.

* Remove the Eigen files.

* Ignore Eigen.

* Remove CMakeFiles.

* Remove CMakeFiles.

* Remove unused files.

* Add all unused files to gitignore.

* Fix gitignore problem.

* Add files in gitignore.

* Add .vs to gitignore.

* Add common/common.hh.

* Fix bug.

* Make project work in VS.

* Make Eigen3 system folder.

* Make the code work in VS.

* Remove jpg using png.

* Add the steps of running on windows.

* add some notes

* Remove the autogeneratd MakeFile.

* Remove the vs.h and move the logic to common.hh

* Recover Makefile.

* Fix the readme problem.

* Disable the jpeg usage in windows.

* Fix the Eigen3 not found problem.

* Fix bug.

* Ignore jpeg in windows.

+++ src/stitch/blender.cc
@@ -42,7 +42,7 @@ Mat32f LinearBlender::run() {
 		memset(weight.ptr(), 0, target_size.y * target_size.x * sizeof(float));
 		fill(target, Color::BLACK);
 #pragma omp parallel for schedule(dynamic)
-		REP(k, images.size()) {
+		REP(k, (int)images.size()) {
 			auto& img = images[k];
 			img.imgref.load();
 			auto& range = img.range;
