+++ src/stitch/multiband.cc
@@ -146,7 +146,7 @@ void MultiBandBlender::create_next_level(int level) {
 	TOTAL_FUNC_TIMER;
 	GaussianBlur blurer(sqrt(level * 2 + 1.0) * 4);	// TODO size
 #pragma omp parallel for schedule(dynamic)
-	REP(i, images.size())
+	REP(i, (int)images.size())
 		next_lvl_images[i].img = blurer.blur(images[i].img);
 }
